Refactor errors: add `InvariantError` to be extended on common hierarchy
