+++ packages/core/src/state/tests/invariants.test.ts
@@ -10,21 +10,27 @@ import {
   removeEntity,
 } from "../mutations";
 import { setTransform } from "../../transform";
+import { EntityNotFoundError } from "../../common/errors";
+import { DuplicateEntityError } from "../errors";
 
-describe("", () => {
+describe("Invariants tests", () => {
   it("Transform은 반드시 존재하는 엔티티에만 연결", () => {
-    const s0 = createEmptyState();
-    const s1 = addEntity(
-      s0,
+    const emptyState =
+      createEmptyState();
+
+    const entityAdded = addEntity(
+      emptyState,
       "a",
       "Box",
     );
 
     expect(() =>
-      setTransform(s1, "b", {
+      setTransform(entityAdded, "b", {
         position: [0, 0, 0],
       }),
-    ).toThrow(/Entity not found: b/);
+    ).toThrow(
+      new EntityNotFoundError("b"),
+    );
   });
 
   it("엔티티 삭제 시 고아 Transform을 남기지 않음", () => {
@@ -68,6 +74,6 @@ describe("", () => {
         "b",
         "Box",
       ),
-    ).toThrow(/Duplicate entity/);
+    ).toThrow(DuplicateEntityError);
   });
 });
