--- packages/core/src/transform/errors.ts
@@ -1,7 +1,0 @@
-export class EntityNotFoundError extends Error {
-  constructor(id: string) {
-    super(`Entity not found: ${id}`);
-
-    this.name = "EntityNotFoundError";
-  }
-}
